requirements.txt
streamlit>=1.32.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=4.1.1
python-weather>=1.0.4
langgraph>=0.0.19
langchain-openai>=0.0.5
langchain>=0.1.0
